clean warnings
