+++ tpls/Dockerfile
@@ -4,7 +4,7 @@ RUN apt-get update
 RUN apt-get install -y git build-essential m4 zlib1g-dev libx11-dev gfortran pkg-config autoconf python3-dev vim tmux nano gdb valgrind libssl-dev
 RUN apt-get autoremove && apt-get clean
 
-RUN git clone https://github.com/goma/goma /opt/goma
+RUN git clone https://github.com/wortiz/goma /opt/goma
 
 RUN PYTHONUNBUFFERED=1 /opt/goma/tpls/install-tpls.py --cc=gcc --cxx=g++ --fc=gfortran --build-shared --download-dir=/tmp/downloads -j 32 --extract-dir=/tmp/extract /opt/goma-libs
